chore: extended demo snapshot, generate script, refresh opts, clippy fixes
- Add generate_extended_demo_snapshot.py (2026-01-01–2026-02-10, ~1845 windows, ~1 MB)
- Snapshot: replace with extended synthetic data; README doc
- refresh_demo_snapshot.sh: START_DATE/END_DATE, MAX_WINDOWS; build --max-windows default 2000
- Clippy: allow too_many_arguments (replay, apply), drop unnecessary casts (policy), is_some_and/is_none_or (validation, cli)

+++ src/state/policy.rs
@@ -57,8 +57,8 @@ impl FeePolicy {
 
     /// Compute operator_share and protocol_fee from gross_spent (integer division).
     pub fn split(&self, gross_spent: u64) -> (u64, u64) {
-        let op = (gross_spent as u64 * self.operator_share_bps as u64) / BPS_MAX as u64;
-        let proto = (gross_spent as u64 * self.protocol_fee_bps as u64) / BPS_MAX as u64;
+        let op = (gross_spent * self.operator_share_bps as u64) / BPS_MAX as u64;
+        let proto = (gross_spent * self.protocol_fee_bps as u64) / BPS_MAX as u64;
         (op, proto)
     }
 }
@@ -96,7 +96,7 @@ impl PolicyConfig {
             ReservePolicy::None => 0,
             ReservePolicy::Fixed { amount } => *amount,
             ReservePolicy::Bps { reserve_bps } => {
-                (gross_spent as u64 * *reserve_bps as u64) / BPS_MAX as u64
+                (gross_spent * *reserve_bps as u64) / BPS_MAX as u64
             }
         }
     }
